Improve getting embedded support
