+++ server/test/mocha.opts
@@ -2,6 +2,7 @@
 --useColors true
 --require source-map-support/register
 --exit
+./dist/embeddedSupport/test
 ./dist/modes/template/test
 ./dist/modes/style/stylus/test
 ./dist/modes/test/
